requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
agate==1.7.1
    # via dbt-core
alembic==1.13.1
    # via dagster
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.14.0
    # via agate
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via dbt-core
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   dagster
    #   dbt-core
    #   dbt-semantic-interfaces
    #   typer
colorama==0.4.6
    # via dbt-core
coloredlogs==14.0
    # via dagster
croniter==2.0.2
    # via dagster
dagster==1.6.9
    # via
    #   dagster-dbt
    #   dagster-duckdb
    #   dagster-duckdb-pandas
dagster-dbt==0.22.9
dagster-duckdb==0.22.9
    # via dagster-duckdb-pandas
dagster-duckdb-pandas==0.22.9
dagster-pipes==1.6.9
    # via dagster
dbt-core==1.7.9
    # via
    #   dagster-dbt
    #   dbt-duckdb
dbt-duckdb==1.7.3
dbt-extractor==0.5.1
    # via dbt-core
dbt-semantic-interfaces==0.4.4
    # via dbt-core
docstring-parser==0.15
    # via dagster
duckdb==0.9.2
    # via
    #   dagster-duckdb
    #   dbt-duckdb
fsspec==2024.2.0
    # via universal-pathlib
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.62.1
    # via
    #   dagster
    #   grpcio-health-checking
grpcio-health-checking==1.62.1
    # via dagster
humanfriendly==10.0
    # via coloredlogs
idna==3.6
    # via
    #   dbt-core
    #   requests
importlib-metadata==6.11.0
    # via dbt-semantic-interfaces
isodate==0.6.1
    # via
    #   agate
    #   dbt-core
jinja2==3.1.3
    # via
    #   dagster
    #   dagster-dbt
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema==4.21.1
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema-specifications==2023.12.1
    # via jsonschema
leather==0.4.0
    # via agate
logbook==1.5.3
    # via dbt-core
mako==1.3.2
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mashumaro==3.12
    # via dbt-core
mdurl==0.1.2
    # via markdown-it-py
minimal-snowplow-tracker==0.0.2
    # via dbt-core
more-itertools==10.2.0
    # via dbt-semantic-interfaces
msgpack==1.0.8
    # via mashumaro
networkx==3.2.1
    # via
    #   dagster-dbt
    #   dbt-core
numpy==1.26.4
    # via pandas
orjson==3.9.15
    # via dagster-dbt
packaging==24.0
    # via
    #   dagster
    #   dagster-dbt
    #   dbt-core
pandas==2.0.3
    # via dagster-duckdb-pandas
parsedatetime==2.6
    # via agate
pathspec==0.11.2
    # via dbt-core
pendulum==3.0.0
    # via dagster
protobuf==4.25.3
    # via
    #   dagster
    #   dbt-core
    #   grpcio-health-checking
pycparser==2.21
    # via cffi
pydantic==2.6.4
    # via
    #   dagster
    #   dbt-semantic-interfaces
pydantic-core==2.16.3
    # via pydantic
pygments==2.17.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   croniter
    #   dagster
    #   dbt-semantic-interfaces
    #   pandas
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via dagster
python-slugify==8.0.4
    # via agate
pytimeparse==1.1.8
    # via agate
pytz==2024.1
    # via
    #   croniter
    #   dagster
    #   dbt-core
    #   pandas
pyyaml==6.0.1
    # via
    #   dagster
    #   dbt-core
    #   dbt-semantic-interfaces
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   dagster
    #   dagster-dbt
    #   dbt-core
    #   minimal-snowplow-tracker
rich==13.7.1
    # via
    #   dagster
    #   dagster-dbt
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
setuptools==69.2.0
    # via dagster
six==1.16.0
    # via
    #   isodate
    #   minimal-snowplow-tracker
    #   python-dateutil
sqlalchemy==2.0.28
    # via
    #   alembic
    #   dagster
sqlparse==0.4.4
    # via dbt-core
structlog==24.1.0
    # via dagster
tabulate==0.9.0
    # via dagster
text-unidecode==1.3
    # via python-slugify
time-machine==2.14.0
    # via pendulum
tomli==2.0.1
    # via dagster
toposort==1.10
    # via dagster
tqdm==4.66.2
    # via dagster
typer==0.9.0
    # via dagster-dbt
typing-extensions==4.10.0
    # via
    #   alembic
    #   dagster
    #   dbt-core
    #   dbt-semantic-interfaces
    #   mashumaro
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
tzdata==2024.1
    # via
    #   pandas
    #   pendulum
universal-pathlib==0.2.2
    # via dagster
urllib3==1.26.18
    # via
    #   dbt-core
    #   requests
watchdog==4.0.0
    # via dagster
zipp==3.18.0
    # via importlib-metadata
